Uploading 3 files
arrivingAnimals.py is the file that tries to comply with the instruction given by our instructor.
arrivingAnimalsReport2.txt is the file that gets written by the arrivingAnimals.py. Though, the instruction for what to have written was vague. I only show my attempt.
animalsLocation.py is a file that was renamed from FileIOSample.py from Week7 repository. Saw questions and attempted them.